Move confirmation logic up
The original logic order came from the jquery implementation that resubmitted the entire request on the dom element when the confirmation promise is resolved. In this version, we call on the internal logic instead of using a new dom event. This means the internal method logic order accuracy is more important and has to be corrected. For example, reprocessing of the request data that is modified via the ajax:confirm-message event.

Fixes #40